Print "checkglobalsi" object

Print method for S3-objects of class "checkglobalsi" characteristic to the "imports" list element of "checkglobals" objects returned by checkglobals, check_pkg or check_source.

Usage

# S3 method for class 'checkglobalsi'
print(x, format = "basic", pattern, ...)

Arguments

x

object inheriting from class "checkglobalsi".

format

character, one of the following two choices:

"basic", (default) prints only the name and source code location of the detected imports.

"detail", prints the name and location of the detected imports, as well as the lines in the source code file comprising the detected imports. The maximum number of lines printed per source code reference can be specified using maxLines.

pattern

an optional regular expression. Only names matching pattern are returned. glob2rx can be used to convert wildcard patterns to regular expressions.

...

additional arguments to configure the printed output. The following arguments can be specified:

all.names, a logical value. If TRUE, all object names are returned. If FALSE, names which begin with a . are omitted. Defaults to TRUE.

includePkgs, a character vector of imported package names. Only the imports for these packages are printed. Defaults to NA, which includes all detected packages in the printed output.

maxRef, the maximum number of printed source code references per detected global/import. Defaults to 1.

maxLines, the maximum number of printed lines per source code reference, only used if format = "detail". Defaults to 5.

use_cli, a logical value indicating if cli should be used to format the printed output. Defaults to TRUE, which means that cli-formatting is attempted if cli is installed.

maxWidth, the maximum column width of the printed output. If cli is installed, the default width is determined by cli::console_width(). If cli is not installed, getOption("width") is checked. If getOption("width") is undefined, the column width defaults to 80.

Value

Returns the object x invisibly (via invisible)

Examples

## R-package from folder
chk <- checkglobals(
  pkg = system.file(
    "unit_tests", "pkg", "testpkg",
    package = "checkglobals"
  )
)
chk$imports
#> 
#> ── Detected imported functions or variables ────────────────────────────────────
#> 
#>  R6         [◼◻◻◻◻◻◻◻◻◻] 1/21
#>   └─ R6Class         functions1.R#139
#>  grid       [◼◻◻◻◻◻◻◻◻◻] 1/21
#>   └─ is.unit         functions1.R#35
#>  methods    [◼◼◼◻◻◻◻◻◻◻] 3/21
#>   ├─ getMethod       functions1.R#73
#>   ├─ setGeneric      functions1.R#54
#>   └─ setMethod       functions1.R#55
#>  parallel   [◼◼◻◻◻◻◻◻◻◻] 2/21
#>   └─ pvec            functions1.R#94 and 1 more...
#>  stats      [◼◼◼◼◼◼◼◻◻◻] 7/21
#>   ├─ aggregate       functions1.R#93
#>   ├─ approxfun       functions1.R#77
#>   ├─ coef            functions1.R#75 and 1 more...
#>   ├─ median          functions1.R#31 and 1 more...
#>   └─ sd              functions1.R#161
#>  stats4     [◼◼◻◻◻◻◻◻◻◻] 2/21
#>   └─ coef            functions1.R#75 and 1 more...
#>  utils      [◼◼◼◼◼◻◻◻◻◻] 5/21
#>   ├─ globalVariables aaa.R#1 and 1 more...
#>   ├─ head            functions1.R#26
#>   ├─ maintainer      aaa.R#8
#>   └─ tail            functions1.R#71

## print selected imports
print(chk, format = "detail", pattern = "coef", which = "import", maxRef = 99)
#> ! Packages required but not installed: pkgA
#> 
#> ── Detected imported functions or variables ────────────────────────────────────
#> 
#>  stats    [◼◼◻◻◻◻◻◻◻◻] 2/4
#>   └─ coef functions1.R#75 and 1 more...
#>  stats4   [◼◼◻◻◻◻◻◻◻◻] 2/4
#>   └─ coef functions1.R#75 and 1 more...
#> 
#> ── Import source code references ───────────────────────────────────────────────
#> 
#> • coef at functions1.R#75:
#>     75: stats::coef(y)
#>   coef at functions1.R#76:
#>     76: stats4::coef(y)

## print without cli-formatting
print(chk$imports, use_cli = FALSE)
#> 
#> ── Detected imported functions or variables ────────────────────────────────────
#> 
#>  R6         [◼◻◻◻◻◻◻◻◻◻] 1/21
#>    • R6Class         functions1.R#139
#>  grid       [◼◻◻◻◻◻◻◻◻◻] 1/21
#>    • is.unit         functions1.R#35
#>  methods    [◼◼◼◻◻◻◻◻◻◻] 3/21
#>    • getMethod       functions1.R#73
#>    • setGeneric      functions1.R#54
#>    • setMethod       functions1.R#55
#>  parallel   [◼◼◻◻◻◻◻◻◻◻] 2/21
#>    • pvec            functions1.R#94 and 1 more...
#>  stats      [◼◼◼◼◼◼◼◻◻◻] 7/21
#>    • aggregate       functions1.R#93
#>    • approxfun       functions1.R#77
#>    • coef            functions1.R#75 and 1 more...
#>    • median          functions1.R#31 and 1 more...
#>    • sd              functions1.R#161
#>  stats4     [◼◼◻◻◻◻◻◻◻◻] 2/21
#>    • coef            functions1.R#75 and 1 more...
#>  utils      [◼◼◼◼◼◻◻◻◻◻] 5/21
#>    • globalVariables aaa.R#1 and 1 more...
#>    • head            functions1.R#26
#>    • maintainer      aaa.R#8
#>    • tail            functions1.R#71